+++ tests/services/sql/parser_test.go
@@ -79,88 +79,6 @@ func TestParseDelete(t *testing.T) {
 	}
 }
 
-func TestParseBatchUpdate(t *testing.T) {
-	parser := mock.NewParser()
-	command := parser.Parse("UPDATE file.txt SET content='a' WHERE content = 'x', SET content='b' WHERE content = 'y'")
-
-	if !command.IsBatch {
-		t.Fatal("expected batch mode")
-	}
-
-	if len(command.Replacements) != 2 {
-		t.Fatalf("expected 2 replacements, got %d", len(command.Replacements))
-	}
-
-	if command.Replacements[0].Replace != "a" {
-		t.Errorf("first replacement: expected 'a', got %s", command.Replacements[0].Replace)
-	}
-
-	if command.Replacements[1].Replace != "b" {
-		t.Errorf("second replacement: expected 'b', got %s", command.Replacements[1].Replace)
-	}
-}
-
-func TestParseBatchDelete(t *testing.T) {
-	parser := mock.NewParser()
-	command := parser.Parse("DELETE FROM file.txt WHERE content = 'x', WHERE content = 'y'")
-
-	if !command.IsBatch {
-		t.Fatal("expected batch mode")
-	}
-
-	if len(command.Deletions) != 2 {
-		t.Fatalf("expected 2 deletions, got %d", len(command.Deletions))
-	}
-}
-
-func TestLikePatternPrefix(t *testing.T) {
-	parser := mock.NewParser()
-	command := parser.Parse("SELECT * FROM f WHERE content LIKE '%test'")
-
-	if !command.Pattern.MatchString("mytest") {
-		t.Error("should match 'mytest'")
-	}
-
-	if command.Pattern.MatchString("testing") {
-		t.Error("should not match 'testing'")
-	}
-}
-
-func TestLikePatternSuffix(t *testing.T) {
-	parser := mock.NewParser()
-	command := parser.Parse("SELECT * FROM f WHERE content LIKE 'test%'")
-
-	if !command.Pattern.MatchString("testing") {
-		t.Error("should match 'testing'")
-	}
-
-	if command.Pattern.MatchString("mytest") {
-		t.Error("should not match 'mytest'")
-	}
-}
-
-func TestLikePatternBoth(t *testing.T) {
-	parser := mock.NewParser()
-	command := parser.Parse("SELECT * FROM f WHERE content LIKE '%test%'")
-
-	if !command.Pattern.MatchString("mytesting") {
-		t.Error("should match 'mytesting'")
-	}
-}
-
-func TestLikePatternExact(t *testing.T) {
-	parser := mock.NewParser()
-	command := parser.Parse("SELECT * FROM f WHERE content LIKE 'test'")
-
-	if !command.Pattern.MatchString("test") {
-		t.Error("should match 'test'")
-	}
-
-	if command.Pattern.MatchString("testing") {
-		t.Error("should not match 'testing' (LIKE without % is exact match)")
-	}
-}
-
 func TestExactMatch(t *testing.T) {
 	parser := mock.NewParser()
 	command := parser.Parse("SELECT * FROM f WHERE content = 'exact'")
